requirements.txt
pysimplegui==4.60.4
websocket==10.4
